refactor(client): split auth_provider into per-concern files (PR 1/7)
First god-module split per the playbook in
.claude/plans/generic-growing-bentley.md. Pure reorganization — zero
behavior change, public API identical.

Layout:
- providers/auth/auth_provider.dart   (358 LoC) — facade: AuthState,
  AuthNotifier shell, login/register/logout/setPresenceStatus, test hooks.
- providers/auth/auth_token_storage.dart (175 LoC, mixin) — Hive +
  SharedPreferences I/O, legacy-credential cleanup, one-shot prefs->secure
  storage migration.
- providers/auth/auth_token_refresh.dart (384 LoC, mixin) — auto-login,
  refresh-coalescing, web cookie + native body paths, the 401-retry
  authenticatedRequest wrapper used by the rest of the app.
- providers/auth_provider.dart (6 LoC) — re-export shim so the ~73
  existing import paths keep working through this PR.

Used the mixin-on-StateNotifier pattern instead of an extension because
StateNotifier.state is @protected — only mixin/subclass instance bodies
can mutate it. Mirrors how conversations_provider.dart already works.

Test moved to test/providers/auth/auth_provider_test.dart to mirror the
new source layout. All 23 existing tests pass unchanged.
